requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.26.165
botocore==1.29.165
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
jmespath==1.0.1
multidict==6.0.4
openai==0.27.5
psycopg2-binary==2.9.6
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2023.6.3
requests==2.31.0
requests-ip-rotator==1.0.14
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.16
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.16
yarl==1.9.2
